requirements.txt
streamlit
numpy
opencv-python-headless
openai
pillow
python-dotenv